Add Packer Terraform code
This adds Packer's Terraform code to the repo, although it does not
activate it (that needs to be done in a separate module).

The Packer code takes in a GCP region and zone, plus an optional list of
firewall subnets (for enabling SSH).

The Packer code created a service account, which is allowed to act as
the Compute Engine default service account.  Cloud Build is also given
permissions to act as the Packer service account (and to generate
tokens, needed by Packer).  The Packer service account has read-only
access to Compute Engine.  It also has write access to Compute Engine
for specific things:

* Managing instances & disks where the name starts with "packer".
* Managing images.
* Managing the packer subnet.

Network-wise, a dedicated VPC and subnet are configured for Packer, with
a firewall rule allowing SSH in from subnets listed in the input
variable.  Also, there is a firewall rule allowing SSH in from all of
Google Cloud, since Cloud Build does not have a defined list of public
IP addresses.