Create a base common Terraform configuration
This is some common Terraform configuration.  It defines base
configuration (for Terraform & Google Cloud), and also sets Google Cloud
Storage as the backend to use for state storage.

This code also locks down default service accounts, and sets Google
Compute Engine project-wide metadata.  That has the side-effect of
blocking all project-wide SSH keys.

Related to #5, but does not close it.