Add a skeleton code directory for gcsgcp code.
This directory tree will hold the Python source for the `gcsgcp-` seties
of commands, which will be available to run on the management nodes and
DTNs.

The Packer code needs to be updated to push this out to the build image.
This directory tree is being set up as a pip-installable package, to
make it easier to integrate into the system.