packer: Add deployment and bootstrapping of code
The new Packer code copies the entire Git tree to the build image.
Although, it might be enough to just copy over the `code` directory,
which has the gcsgcp source.

The new Packer code also copies and executes a bootstrap script.  The
bootstrap script uses the system python3 (currently 3.7) to make a venv,
which is then updated and used to build & install the gcsgcp code.
Finally, a profile shell script is put in place that activates the venv
on any new interactive login sessions, so that the gcsgcp commands are
available for use directly by admins.