packer: Change _PACKER_DIR default to `packer`
Originally, the `_PACKER_DIR` Cloud Build substition was set to `.`.
But now the build accesses files elsewhere in the repository.  Cloud
Build was already uploading the entire repository; now we update
`_PACKER_DIR` to make `packer` the default, and we update the
instructions for manual builds to not change the substition.