terraform: Start adding code for GCS
This is the start of the Terraform code to deploy Globus Connect server
instances.

First up is the common code, enabling services that we'll need, and
creating a new network for GCS stuff.

We also define a sub-module, which will be used to configure common
permissions that we'll need for different service accounts.