terraform: Add code for GCS management nodes
The management node has a number of things:

* A service account, with permissions needed for the VM to run.  This is
where the common configuration is applied.  It gives access to write to
logs and monitoring and traces, and to access the servicecontrol and
service management APIs.

* The service account is given read-only access to Google Compute, and
it's given OS Login admin access to log in to the DTNs.  The idea (to be
tested) is that this could be used to SSH from management node to DTN.

* A new subnet is made for the management nodes.

* A NAT gateway (and required Cloud Router) is created, with the NAT
gateway set up to only accept traffic from systems in the
management-node subnet, from VMs where the primary IP is in that subnet.
Note that the timeouts for USP sessions and established TCP sessions are
raised to 5 minutes.  Also, full logging of NAT sessions is enabled.

* A firewall rule is put in place to allow SSH in through the
Identity-Aware Proxy.

Finally, the actual instance template!  It's an e2-micro, using the new
service account (and scopes to match).  It has a single subnet with no
public IP, tagged to allow SSH through IAP.  OS Login is also enabled.
It's not preemptible, and it allows migration and auto-restart.  Secure
Boot is enabled; Confidential Compute is not.  It uses the GCS image as
the boot disk.  Client ID and node type are set in metadata.