terraform: Enable GCS via module option
This change to Terraform common code conditionally enables the new GCS
code via a module, with a count of either 1 or 0.

The GCS code cannot be run immediately, because it requires at least one
image be in the `globus` family.  That won't be possible until Cloud
Build runs at least once, which won't be possible until the `packer`
Terraform code has run.  So, we have a variable to control if the GCS
module should be enabled.

We also add client_id as a top-level variable, since the GCS images need
it in instance metadata.