packer: Move the package cleanup earlier in the build.
As part of package cleanup, we do an `apt-get autoremove`.  This is
meant to remove packages that were installed as dependencies of packages
that have since been removed.

We need to do this before we do the bootstrap work.  That is because we
might autoremove Python packages, and depending on which Python packages
are removed, that could cause breakage.