packer_builder: Switch base image to Google Cloud SDK
Some upcoming changes to the Packer code will use the `gcloud` and
`gsutil` commands.  But right now, the Packer container does not have
those commands installed.  So, I am switching the base container image
to the Google Cloud SDK image
(https://cloud.google.com/sdk/docs/downloads-docker).

This image uses Debian as a base, which is close enough to our normal
(Ubuntu) base image that it's not a problem to switch.  The image in
question is maintained by Google, so I don't think it will go
unmaintained anytime soon.