packer: Upload build artifacts to Cloud Storage
At the end of the Packer build, but before the image is created,
generate lists of Debian and Pip packages that were installed, copy them
out of the instance, and upload them to Cloud Storage for
record-keeping.

This uses the `gsutil` command from the Google Cloud SDK container
image.

The README is also updated, to note the need for a Cloud Storage bucket,
and appropriate permissions for the Cloud Build Default Service Account.