terraform/packer: Add Pub/Sub Topic
Created a new Pub/Sub Topic, which will receive notifications when new
Compute Engine images are created.

In addition to creating the topic, a schema is set: Messages must be
a JSON object containing one key, `image_name`, which contains the name
of the newly-created Compute Engine image.

Finally, the Compute Engine Default Service Account is given explicit
permission to post to the topic.