packer: Post Pub/Sub notification on completion
When Packer is done building a new Compute Engine image, post a message to a
Pub/Sub Topic.  The message is a JSON object, containing the key `image_name`,
whose value is the name of the just-created Compute Engine image.

This takes advantage of the new Google Cloud SDK container image, that we are
using as the base for the Packer image.  `gcloud` is called to publish the
message.

For this to work, a new Packer variable, `pubsub_topic`, is required: This is
the name of the Pub/Sub Topic.  Similarly, a new `_PUBSUB_TOPIC` substitution
is created, which does *not* have a default.