packer: Switch timestamp to interpolate immediately.
The original timestamp was using `{{timestamp}}`.  Turns out, that isn't
actually evaluated into a timestamp until the time when it actually used
somewhere.  In other words, if you use `{{timestamp}}` in a local's
expression, it won't be interpolated immediately.  This meant that the
timestamp used for the image name was not the same as the timestamp used
for the artifact upload.

This means switching to something else.  `build_timestamp` is defined as
`timestamp()`, which is interpolated at the time it is defined.
`build_time` is updated to use that (it previously included
`timestamp()` directly), and `image_name` is redefined to use the new
`build_timestamp`.

Unfortunately, `formatdate` does not have a way to request a 'seconds
since Unix epoch', so the image name format is changing.  Code which
sorts image names will need to be checked.