terraform/packer: Add Cloud Build triggers
This adds a Cloud Build trigger, and sets up for automatically invoking
said trigger.

The first trigger created is one which runs when a push is made to the
GitHub repo's main branch.

The second trigger, which is not actually created (see the NOTE below),
is triggered when a message is sent to a Pub/Sub Topic.  The Topic is
created here, and a Cloud Scheduiler job is created to send a message to
the Topic on a regular schedule.  Through this, automated builds are
achieved.

NOTE: At this time, it is not possible for Terraform to create a Cloud
Build Trigger which is triggered by Pub/Sub *and* which pulls from
GitHub.  Until that is available, one Cloud Build Trigger will need to
be created manually.