terraform/packer: Define Functions for Slack and Cleanup
This code defines two Cloud Functions, each of which is triggered by a
post to a particlar Pub/Sub Topic:

One Function is triggered whenever a post is made to the image_updated
Topic.  The posted message contains the name of the just-built image,
and the Function uses this to clean up old images and make diffs.

The other Function is triggered whenever a post is made to the
slack_message Topic.  The posted message JSON is POSTed directly to a
Slack webhook.

This code also defines a Secret Manager Secret containing the webhook
URL, and populates the Secret with the webhook URL provided as an input
variable.

NOTE: Before Terraform is run, the admin will need to `cd` into each of
the sub-directories under `functions`, and run `make dist` to create the
Zip files that will be uploaded.

WARNING: Right now, the code Zip files are stored in a Cloud Storage
bucket which has a purge policy.  This means that either there need to
be regular updates of the code, or the policy needs to be changed.