terraform/gcs: Allow posting to Slack from instances
Give our instances (DTNs and Management nodes) access to post Slack
messages.  Permissions are granted to publish to the Slack Pub/Sub
Topic, and the Topic ID is sent through to the instances as metadata.

Since this is a common need for both instance types, the actual
permissions enabling this are in sa_common.