terraform/gcs: Set NAT ports per VM to 64
This explicitly sets the minimum number of NAT ports per NATed VM to 64.
This is actually the default value, which is used when you explicity set
a value.  The reason we set it here, is because if you try leaving the
field blank in the web UI, the value 64 is explicitly set, and that
makes it into the configuation (as an explicit value).  That causes
confusion with Terraform, which expects the value to be 64.  So every
time you run Terraform after that, it tries to change the value from 64
to `(null)`.

Explicitly setting the value to 64 stops Terraform from constantly
trying to change it.