code: Include system packages in venv
In commit 0a3ea19, the systemd Python package is installed.  The package
is installed into the system package space.  That's fine, since we're
using a venv initialized from system Python, but by default the venv
won't search system packages.

This changes the venv to allow accessing system packages.  One possible
side-effect is that, if we request installation (or depend) on a package
that's present on the system, `pip` might not install it.  The next
build will show the effects of this.