packer: Switch from Cloud Monitoring to Cloud Ops
In the past, Google had two agents for 'observability': The Cloud
Monitoring agent essentially wraps around collectd, collecting metrics
from the OS (like memory used %) and sending them to the Stackdriver
Monitoring API.

There was another agent, the Logging agent, which would do the same
thing for logs.  Now, both things are in one agent.  So, let's switch to
that!