packer: Change auditd rules upload and copy process
This makes two changes:

First, after a more-careful reading of
http://www.packer.io/docs/provisioners/file#directory-uploads, I
realized the trailing slash from the source path needs to be removed.  I
want the entire `auditd_rules` directory to be uploaded, not just the
contents of the directory.  I also removed the trailing slash from the
destination, since the linked docs don't use a trailing slash with the
destination.

Finally, instead of doing a `mv` to move the files into place, a `cp` is
used.  This is because `mv` tries to preserve things like ownership,
while `cp` (by default) does not.